admin/pyproject.toml
[tool.poetry]
name = "voltaserve-admin"
version = "0.0.1"
description = "Asynchronous API for voltaserve-admin microservice"
authors = ["Piotr Łoboda <voltaserve@lobson.pl>"]
maintainers = ["Piotr Łoboda <voltaserve@lobson.pl>"]
packages = [
        {include = 'api/'},
        {include = 'api/routers'},
        {include = 'api/models'},
        {include = 'api/database'}
    ]

classifiers = [
        "Programming Language :: Python :: 3.12",
        "Framework :: FastAPI",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: POSIX :: Linux",
        "Development Status :: 3 - Alpha"
    ]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "2.3.4"
PyJWT = "^2.8.0"
websockets = "12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
